project/requirements.txt
beautifulsoup4
flask
gunicorn
jieba
line-bot-sdk
pymongo[srv]
python-dotenv
pytz
rich
uwsgi